Get rid of overlapping transaction
publishAsync needs a commited ExtensionVersion
