+++ server/src/test/java/org/eclipse/openvsx/RegistryAPITest.java
@@ -50,6 +50,7 @@ import org.eclipse.openvsx.json.SearchEntryJson;
 import org.eclipse.openvsx.json.SearchResultJson;
 import org.eclipse.openvsx.json.UserJson;
 import org.eclipse.openvsx.publish.PublishExtensionVersionHandler;
+import org.eclipse.openvsx.publish.PublishExtensionVersionService;
 import org.eclipse.openvsx.repositories.RepositoryService;
 import org.eclipse.openvsx.search.ExtensionSearch;
 import org.eclipse.openvsx.search.ISearchService;
@@ -86,7 +87,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 @MockBean({
     ClientRegistrationRepository.class, UpstreamRegistryService.class, GoogleCloudStorageService.class,
     AzureBlobStorageService.class, VSCodeIdService.class, DownloadCountService.class, AzureDownloadCountService.class,
-    LockProvider.class, CacheService.class, EclipseService.class, PublishExtensionVersionHandler.class
+    LockProvider.class, CacheService.class, EclipseService.class, PublishExtensionVersionService.class
 })
 public class RegistryAPITest {
 
@@ -1744,6 +1745,11 @@ public class RegistryAPITest {
         @Bean
         LatestExtensionVersionDTOCacheKeyGenerator latestExtensionVersionDTOCacheKeyGenerator() {
             return new LatestExtensionVersionDTOCacheKeyGenerator();
+        }
+
+        @Bean
+        PublishExtensionVersionHandler publishExtensionVersionHandler() {
+            return new PublishExtensionVersionHandler();
         }
     }
 }
